// frontend/src/modules/shared/types/style.types.ts
export interface AppTheme {
  primaryColor: string
  primaryHover: string
  primaryActive: string
  secondaryColor: string
  primaryBackgroundColor: string
  inputBackground: string
  inputSelectedBackground: string
  tableTitle: string
  primaryRadius: string
  inputStroke: string
  inputPlaceholder: string
  textColor: string
  textDarker: string
  yellowColor: string
  orangeColor: string
  successColor: string
  calendulaGold: string
  dangerColor: string
  purpleColor: string
  popupBackground: string
  modalBackground: string
  ticketColor: string
  borderDark: string
  transition: string
  whiteColor: string
}